Refactor app config into modular components
