+++ src/core/config/app_config.py
@@ -3,86 +3,46 @@ from __future__ import annotations
 import json
 import logging
 import os
-from collections.abc import Callable, Mapping
-from enum import Enum
+from collections.abc import Mapping
 from pathlib import Path
 from typing import Any, cast
 
-from pydantic import ConfigDict, Field, field_validator, model_validator
+from pydantic import ConfigDict, Field
 
+from src.core.config._app_config_models import (
+    AuthConfig,
+    BackendConfig,
+    BackendSettings,
+    BruteForceProtectionConfig,
+    EditPrecisionConfig,
+    EmptyResponseConfig,
+    LogLevel,
+    LoggingConfig,
+    ModelAliasRule,
+    PlanningPhaseConfig,
+    RewritingConfig,
+    SessionConfig,
+    SessionContinuityConfig,
+    ToolCallReactorConfig,
+)
+from src.core.config._app_config_utils import (
+    _collect_api_keys_from_env,
+    _env_to_bool,
+    _env_to_float,
+    _env_to_int,
+    _flatten_dict,
+    _get_api_keys_from_env,
+    _get_by_path,
+    _get_env_value,
+    _merge_dicts,
+    _process_api_keys,
+    _set_by_path,
+    _to_float,
+    _to_int,
+    get_openrouter_headers,
+    load_config,
+)
 from src.core.config.parameter_resolution import ParameterResolution, ParameterSource
-
-
-def get_openrouter_headers(cfg: dict[str, str], api_key: str) -> dict[str, str]:
-    """Construct headers for OpenRouter requests.
-
-    Be tolerant of minimal cfg dicts provided by tests by falling back to
-    sensible defaults when optional keys are absent.
-    """
-    referer: str = cfg.get("app_site_url", "http://localhost:8000")
-    x_title: str = cfg.get("app_x_title", "InterceptorProxy")
-    return {
-        "Authorization": f"Bearer {api_key}",
-        "Content-Type": "application/json",
-        "HTTP-Referer": referer,
-        "X-Title": x_title,
-    }
-
-
-def _collect_api_keys_from_env(
-    base_name: str,
-    env: Mapping[str, str],
-    resolution: ParameterResolution | None = None,
-) -> dict[str, str]:
-    """Collect API keys from environment variables with parameter resolution tracking."""
-    single_key = env.get(base_name)
-    numbered_keys: dict[str, str] = {}
-    numbered_key_names = []
-    for i in range(1, 21):
-        key_name = f"{base_name}_{i}"
-        key = env.get(key_name)
-        if key:
-            numbered_keys[key_name] = key
-            numbered_key_names.append(key_name)
-
-    if single_key and numbered_keys:
-        logger.warning(
-            "Both %s and %s_<n> environment variables are set. Prioritizing %s_<n> and ignoring %s.",
-            base_name,
-            base_name,
-            base_name,
-            base_name,
-        )
-        if resolution is not None:
-            resolution.record(
-                f"backends.{base_name.lower().replace('_', '')}.api_key",
-                list(numbered_keys.values()),
-                ParameterSource.ENVIRONMENT,
-                origin=",".join(numbered_key_names),
-            )
-        return numbered_keys
-
-    if single_key:
-        result = {base_name: single_key}
-        if resolution is not None:
-            resolution.record(
-                f"backends.{base_name.lower().replace('_', '')}.api_key",
-                list(result.values()),
-                ParameterSource.ENVIRONMENT,
-                origin=base_name,
-            )
-        return result
-
-    if resolution is not None and numbered_keys:
-        resolution.record(
-            f"backends.{base_name.lower().replace('_', '')}.api_key",
-            list(numbered_keys.values()),
-            ParameterSource.ENVIRONMENT,
-            origin=",".join(numbered_key_names),
-        )
-    return numbered_keys
-
-
 from src.core.domain.configuration.app_identity_config import AppIdentityConfig
 from src.core.domain.configuration.assessment_config import AssessmentConfig
 from src.core.domain.configuration.header_config import (
@@ -95,654 +55,10 @@ from src.core.domain.configuration.reasoning_aliases_config import (
 from src.core.domain.configuration.sandboxing_config import SandboxingConfiguration
 from src.core.interfaces.configuration_interface import IConfig
 from src.core.interfaces.model_bases import DomainModel
-
-# Note: Avoid self-imports to prevent circular dependencies. Classes are defined below.
+from src.core.services.backend_registry import backend_registry
 
 logger = logging.getLogger(__name__)
-
-
-def _process_api_keys(keys_string: str) -> list[str]:
-    """Process a comma-separated string of API keys."""
-    keys = keys_string.split(",")
-    result: list[str] = []
-    for key in keys:
-        stripped_key = key.strip()
-        if stripped_key:
-            result.append(stripped_key)
-    return result
-
-
-def _get_api_keys_from_env(
-    env: Mapping[str, str], resolution: ParameterResolution | None = None
-) -> list[str]:
-    """Get API keys from environment variables."""
-    result: list[str] = []
-
-    # Get API keys from API_KEYS environment variable
-    api_keys_raw: str | None = env.get("API_KEYS")
-    if api_keys_raw and isinstance(api_keys_raw, str):
-        result.extend(_process_api_keys(api_keys_raw))
-
-    if result and resolution is not None:
-        resolution.record(
-            "auth.api_keys",
-            result,
-            ParameterSource.ENVIRONMENT,
-            origin="API_KEYS",
-        )
-
-    return result
-
-
-def _env_to_bool(
-    name: str,
-    default: bool,
-    env: Mapping[str, str],
-    *,
-    path: str | None = None,
-    resolution: ParameterResolution | None = None,
-) -> bool:
-    """Return an environment variable parsed as a boolean flag."""
-    value = env.get(name)
-    if value is None:
-        return default
-    result = value.strip().lower() in {"1", "true", "yes", "on"}
-    if resolution is not None and path is not None:
-        resolution.record(path, result, ParameterSource.ENVIRONMENT, origin=name)
-    return result
-
-
-def _env_to_int(
-    name: str,
-    default: int,
-    env: Mapping[str, str],
-    *,
-    path: str | None = None,
-    resolution: ParameterResolution | None = None,
-) -> int:
-    """Return an environment variable parsed as an integer."""
-    value = env.get(name)
-    if value is None:
-        return default
-    try:
-        result = int(value)
-    except (TypeError, ValueError):
-        result = default
-    if resolution is not None and path is not None and value is not None:
-        resolution.record(path, result, ParameterSource.ENVIRONMENT, origin=name)
-    return result
-
-
-def _env_to_float(
-    name: str,
-    default: float,
-    env: Mapping[str, str],
-    *,
-    path: str | None = None,
-    resolution: ParameterResolution | None = None,
-) -> float:
-    """Return an environment variable parsed as a float."""
-    value = env.get(name)
-    if value is None:
-        return default
-    try:
-        result = float(value)
-    except (TypeError, ValueError):
-        result = default
-    if resolution is not None and path is not None and value is not None:
-        resolution.record(path, result, ParameterSource.ENVIRONMENT, origin=name)
-    return result
-
-
-def _get_env_value(
-    env: Mapping[str, str],
-    name: str,
-    default: Any,
-    *,
-    path: str,
-    resolution: ParameterResolution | None = None,
-    transform: Callable[[str], Any] | None = None,
-) -> Any:
-    """Return an environment variable value and optionally record its source."""
-
-    if name in env:
-        raw_value = env[name]
-        value = transform(raw_value) if transform is not None else raw_value
-        if resolution is not None:
-            resolution.record(path, value, ParameterSource.ENVIRONMENT, origin=name)
-        return value
-    return default
-
-
-def _to_int(value: str, fallback: int) -> int:
-    try:
-        return int(value)
-    except (TypeError, ValueError):
-        return fallback
-
-
-def _to_float(value: str, fallback: float | None) -> float | None:
-    try:
-        return float(value)
-    except (TypeError, ValueError):
-        return fallback
-
-
-class LogLevel(str, Enum):
-    """Log levels for configuration."""
-
-    DEBUG = "DEBUG"
-    INFO = "INFO"
-    WARNING = "WARNING"
-    ERROR = "ERROR"
-    CRITICAL = "CRITICAL"
-
-
-class BackendConfig(DomainModel):
-    """Configuration for a backend service."""
-
-    model_config = ConfigDict(frozen=True)
-
-    api_key: list[str] = Field(default_factory=list)
-    api_url: str | None = None
-    models: list[str] = Field(default_factory=list)
-    timeout: int = 120  # seconds
-    identity: AppIdentityConfig | None = None
-    extra: dict[str, Any] = Field(default_factory=dict)
-
-    @field_validator("api_key", mode="before")
-    @classmethod
-    def validate_api_key(cls, v: Any) -> list[str]:
-        """Ensure api_key is always a list."""
-        if isinstance(v, str):
-            return [v]
-        return v if isinstance(v, list) else []
-
-    @field_validator("api_url")
-    @classmethod
-    def validate_api_url(cls, v: str | None) -> str | None:
-        """Validate the API URL if provided."""
-        if v is not None and not v.startswith(("http://", "https://")):
-            raise ValueError("API URL must start with http:// or https://")
-        return v
-
-
-class AuthConfig(DomainModel):
-    """Authentication configuration."""
-
-    model_config = ConfigDict(frozen=True)
-
-    disable_auth: bool = False
-    api_keys: list[str] = Field(default_factory=list)
-    auth_token: str | None = None
-    redact_api_keys_in_prompts: bool = True
-    trusted_ips: list[str] = Field(default_factory=list)
-    brute_force_protection: BruteForceProtectionConfig = Field(
-        default_factory=lambda: BruteForceProtectionConfig()
-    )
-
-
-class BruteForceProtectionConfig(DomainModel):
-    """Configuration for brute-force protection on API authentication."""
-
-    model_config = ConfigDict(frozen=True)
-
-    enabled: bool = True
-    max_failed_attempts: int = 5
-    ttl_seconds: int = 900
-    initial_block_seconds: int = 30
-    block_multiplier: float = 2.0
-    max_block_seconds: int = 3600
-
-
-class LoggingConfig(DomainModel):
-    """Logging configuration."""
-
-    model_config = ConfigDict(frozen=True)
-
-    level: LogLevel = LogLevel.INFO
-    request_logging: bool = False
-    response_logging: bool = False
-    log_file: str | None = None
-    # Optional separate wire-capture log file; when set, all outbound requests
-    # and inbound replies/SSE payloads are captured verbatim to this file.
-    capture_file: str | None = None
-    # Optional max size in bytes; when exceeded, rotate current capture to
-    # `<capture_file>.1` and start a new file (overwrite existing .1).
-    capture_max_bytes: int | None = None
-    # Optional per-chunk truncation size in bytes for streaming capture. When
-    # set, stream chunks written to capture are truncated to this size with a
-    # short marker appended; streaming to client remains unmodified.
-    capture_truncate_bytes: int | None = None
-    # Optional number of rotated files to keep (e.g., file.1..file.N). If not
-    # set or <= 0, keeps a single rotation (file.1). Used only when
-    # capture_max_bytes is set.
-    capture_max_files: int | None = None
-    # Time-based rotation period in seconds (default 1 day). If set <= 0, time
-    # rotation is disabled.
-    capture_rotate_interval_seconds: int = 86400
-    # Total disk cap across current capture file and rotated files. If set <= 0,
-    # disabled. Default is 100 MiB.
-    capture_total_max_bytes: int = 104857600
-    # Buffer size for wire capture writes (bytes). Default 64KB.
-    capture_buffer_size: int = 65536
-    # How often to flush buffer to disk (seconds). Default 1.0 second.
-    capture_flush_interval: float = 1.0
-    # Maximum entries to buffer before forcing flush. Default 100.
-    capture_max_entries_per_flush: int = 100
-
-
-class ToolCallReactorConfig(DomainModel):
-    """Configuration for the Tool Call Reactor system.
-
-    The Tool Call Reactor provides event-driven reactions to tool calls
-    from LLMs, allowing custom handlers to monitor, modify, or replace responses.
-    """
-
-    model_config = ConfigDict(frozen=True)
-
-    enabled: bool = True
-    """Whether the Tool Call Reactor is enabled."""
-
-    apply_diff_steering_enabled: bool = True
-    """Whether the legacy apply_diff steering handler is enabled."""
-
-    apply_diff_steering_rate_limit_seconds: int = 60
-    """Legacy rate limit window for apply_diff steering in seconds.
-
-    Controls how often steering messages are shown for apply_diff tool calls
-    within the same session. Default: 60 seconds (1 message per minute).
-    """
-
-    apply_diff_steering_message: str | None = None
-    """Legacy custom steering message for apply_diff tool calls.
-
-    If None, uses the default message. Can be customized to fit your workflow.
-    """
-
-    pytest_full_suite_steering_enabled: bool = False
-    """Whether steering for full pytest suite commands is enabled."""
-
-    pytest_full_suite_steering_message: str | None = None
-    """Optional custom steering message when detecting full pytest suite runs."""
-
-    pytest_context_saving_enabled: bool = False
-    """Whether pytest context-saving command rewrites are enabled."""
-
-    fix_think_tags_enabled: bool = False
-    """Whether correction of improperly formatted <think> tags is enabled."""
-
-    # New: fully configurable steering rules
-    steering_rules: list[dict[str, Any]] = Field(default_factory=list)
-    """Configurable steering rules.
-
-    Each rule is a dict describing when to trigger steering and what message to
-    return. See README for details. Minimal fields:
-      - name: Unique rule name
-      - enabled: bool
-      - triggers: { tool_names: [..], phrases: [..] }
-      - message: Replacement content when swallowed
-      - rate_limit: { calls_per_window: int, window_seconds: int }
-      - priority: int (optional; higher runs first)
-    """
-
-    # Tool access control policies
-    access_policies: list[dict[str, Any]] = Field(default_factory=list)
-    """Tool access control policies.
-
-    Each policy defines which tools are allowed or blocked for specific models/agents.
-    Minimal fields:
-      - name: Unique policy identifier
-      - model_pattern: Regex pattern for matching model names
-      - default_policy: "allow" or "deny"
-    Optional fields:
-      - agent_pattern: Regex pattern for matching agents
-      - allowed_patterns: List of regex patterns for allowed tools
-      - blocked_patterns: List of regex patterns for blocked tools
-      - block_message: Message to return when blocking a tool call
-      - priority: int (higher values take precedence)
-    """
-
-
-class PlanningPhaseConfig(DomainModel):
-    """Configuration for planning phase model routing."""
-
-    model_config = ConfigDict(frozen=True)
-
-    enabled: bool = False
-    strong_model: str | None = None
-    max_turns: int = 10
-    max_file_writes: int = 1
-    # Optional parameter overrides for the strong model
-    overrides: dict[str, Any] | None = None
-
-
-class SessionContinuityConfig(DomainModel):
-    """Configuration for intelligent session continuity detection."""
-
-    model_config = ConfigDict(frozen=True)
-
-    enabled: bool = True
-    fuzzy_matching: bool = True
-    max_session_age_seconds: int = 604800  # 7 days
-    fingerprint_message_count: int = 5
-    client_key_includes_ip: bool = True
-
-
-class SessionConfig(DomainModel):
-    """Session management configuration."""
-
-    model_config = ConfigDict(frozen=True)
-
-    cleanup_enabled: bool = True
-    cleanup_interval: int = 3600  # 1 hour
-    max_age: int = 86400  # 1 day
-    default_interactive_mode: bool = True
-    force_set_project: bool = False
-    disable_interactive_commands: bool = False
-    project_dir_resolution_model: str | None = None
-    project_dir_resolution_mode: str = "hybrid"
-    tool_call_repair_enabled: bool = True
-    # Max per-session buffer for tool-call repair streaming (bytes)
-    tool_call_repair_buffer_cap_bytes: int = 64 * 1024
-    json_repair_enabled: bool = True
-    # Max per-session buffer for JSON repair streaming (bytes)
-    json_repair_buffer_cap_bytes: int = 64 * 1024
-    json_repair_strict_mode: bool = False
-    json_repair_schema: dict[str, Any] | None = None  # Added
-    tool_call_reactor: ToolCallReactorConfig = Field(
-        default_factory=ToolCallReactorConfig
-    )
-    dangerous_command_prevention_enabled: bool = True
-    dangerous_command_steering_message: str | None = None
-    pytest_compression_enabled: bool = True
-    pytest_compression_min_lines: int = 30
-    pytest_full_suite_steering_enabled: bool | None = None
-    pytest_full_suite_steering_message: str | None = None
-    fix_think_tags_enabled: bool = False
-    fix_think_tags_streaming_buffer_size: int = 4096
-    planning_phase: PlanningPhaseConfig = Field(default_factory=PlanningPhaseConfig)
-    max_per_session_backends: int = 32
-    session_continuity: SessionContinuityConfig = Field(
-        default_factory=SessionContinuityConfig
-    )
-    tool_access_global_overrides: dict[str, Any] | None = None
-    # Tool call processing behavior configuration
-    force_reprocess_tool_calls: bool = False
-    log_skipped_tool_calls: bool = False
-
-    @model_validator(mode="before")
-    @classmethod
-    def _sync_pytest_full_suite_settings(cls, values: dict[str, Any]) -> dict[str, Any]:
-        """Keep pytest full-suite steering settings mirrored with reactor config."""
-        reactor_config = values.get("tool_call_reactor")
-
-        # Convert to dict if it's already a ToolCallReactorConfig instance
-        if isinstance(reactor_config, ToolCallReactorConfig):
-            reactor_config_dict = reactor_config.model_dump()
-        elif isinstance(reactor_config, dict):
-            reactor_config_dict = dict(reactor_config)
-        else:
-            reactor_config_dict = {}
-
-        enabled = values.get("pytest_full_suite_steering_enabled")
-        message = values.get("pytest_full_suite_steering_message")
-
-        # Update the dict instead of mutating frozen model
-        if enabled is not None:
-            reactor_config_dict["pytest_full_suite_steering_enabled"] = enabled
-        else:
-            values["pytest_full_suite_steering_enabled"] = reactor_config_dict.get(
-                "pytest_full_suite_steering_enabled", False
-            )
-
-        if message is not None:
-            reactor_config_dict["pytest_full_suite_steering_message"] = message
-        else:
-            values["pytest_full_suite_steering_message"] = reactor_config_dict.get(
-                "pytest_full_suite_steering_message"
-            )
-
-        # Store the dict - Pydantic will convert it to ToolCallReactorConfig
-        values["tool_call_reactor"] = reactor_config_dict
-        return values
-
-
-class EmptyResponseConfig(DomainModel):
-    """Configuration for empty response handling."""
-
-    model_config = ConfigDict(frozen=True)
-
-    enabled: bool = True
-    """Whether the empty response recovery is enabled."""
-
-    max_retries: int = 1
-    """Maximum number of retries for empty responses."""
-
-
-class ModelAliasRule(DomainModel):
-    """A rule for rewriting a model name."""
-
-    model_config = ConfigDict(frozen=True)
-
-    pattern: str
-    replacement: str
-
-
-class RewritingConfig(DomainModel):
-    """Configuration for content rewriting."""
-
-    model_config = ConfigDict(frozen=True)
-
-    enabled: bool = False
-    config_path: str = "config/replacements"
-
-
-class EditPrecisionConfig(DomainModel):
-    """Configuration for automated edit-precision tuning.
-
-    When enabled, detects agent edit-failure prompts and lowers sampling
-    parameters for the next single call to improve precision.
-    """
-
-    model_config = ConfigDict(frozen=True)
-
-    enabled: bool = True
-    temperature: float = 0.1
-    # Only applied if override_top_p is True; otherwise top_p remains unchanged
-    min_top_p: float | None = 0.3
-    # Control whether top_p/top_k are overridden by this feature
-    override_top_p: bool = False
-    override_top_k: bool = False
-    # Target top_k to apply when override_top_k is True (for providers that support it, e.g., Gemini)
-    target_top_k: int | None = None
-    # Optional regex pattern; when set, agents with names matching this pattern
-    # will be excluded (feature disabled) even if enabled=True.
-    exclude_agents_regex: str | None = None
-
-
-from src.core.services.backend_registry import (
-    backend_registry,  # Updated import path
-)
-
-
-class BackendSettings(DomainModel):
-    """Settings for all backends.
-
-    Note: This class is intentionally not frozen because it needs to support
-    dynamic backend configurations that are added at runtime. Backend configs
-    are stored in __dict__ to allow attribute-style access (e.g., config.backends.openai)
-    without pre-defining all possible backends as fields.
-    """
-
-    model_config = ConfigDict(frozen=False, extra="allow")
-
-    default_backend: str = "openai"
-    static_route: str | None = (
-        None  # Force all requests to backend:model (e.g., "gemini-oauth-plan:gemini-2.5-pro")
-    )
-    disable_gemini_oauth_fallback: bool = False
-    disable_hybrid_backend: bool = False
-    hybrid_backend_repeat_messages: bool = False
-    reasoning_injection_probability: float = Field(
-        default=1.0,
-        ge=0.0,
-        le=1.0,
-        description="Probability of using the reasoning model for a request in the hybrid backend.",
-    )
-
-    def __init__(self, **data: Any) -> None:
-        # Separate standard fields from backend-specific configs
-        known_fields = set(self.model_fields.keys())
-
-        init_data = {k: v for k, v in data.items() if k in known_fields}
-        backend_data = {k: v for k, v in data.items() if k not in known_fields}
-
-        # Initialize the model with standard fields
-        super().__init__(**init_data)
-
-        # Manually set the backend configurations
-        for backend_name, config_data in backend_data.items():
-            if isinstance(config_data, dict):
-                self.__dict__[backend_name] = BackendConfig(**config_data)
-            elif isinstance(config_data, BackendConfig):
-                self.__dict__[backend_name] = config_data
-
-        # Ensure all registered backends have a config
-        for backend_name in backend_registry.get_registered_backends():
-            if backend_name not in self.__dict__:
-                self.__dict__[backend_name] = BackendConfig()
-
-        self._initialization_complete = True
-
-    def __getitem__(self, key: str) -> BackendConfig:
-        """Allow dictionary-style access to backend configs."""
-        if key in self.__dict__:
-            return cast(BackendConfig, self.__dict__[key])
-        raise KeyError(f"Backend '{key}' not found")
-
-    def __setitem__(self, key: str, value: BackendConfig) -> None:
-        """Allow dictionary-style setting of backend configs."""
-        self.__dict__[key] = value
-
-    def __setattr__(self, name: str, value: Any) -> None:
-        """Allow attribute-style assignment for backend configs."""
-        if (
-            name in {"default_backend"}
-            or name.startswith("_")
-            or name in self.model_fields
-        ):
-            super().__setattr__(name, value)
-            return
-        if isinstance(value, BackendConfig):
-            config = value
-        elif isinstance(value, dict):
-            config = BackendConfig(**value)
-        else:
-            config = BackendConfig()
-        self.__dict__[name] = config
-
-    def get(self, key: str, default: Any = None) -> Any:
-        """Dictionary-style get with default."""
-        return cast(BackendConfig | None, self.__dict__.get(key, default))
-
-    @property
-    def functional_backends(self) -> set[str]:
-        """Get the set of functional backends (those with API keys)."""
-        functional: set[str] = set()
-        registered = backend_registry.get_registered_backends()
-        for backend_name in registered:
-            if backend_name in self.__dict__:
-                config: Any = self.__dict__[backend_name]
-                if isinstance(config, BackendConfig) and config.api_key:
-                    functional.add(backend_name)
-
-        # Consider OAuth-style backends functional even without an api_key in config,
-        # since they source credentials from local auth stores (e.g., CLI-managed files).
-        oauth_like: set[str] = set()
-        for name in registered:
-            if name.endswith("-oauth") or name.startswith("gemini-oauth"):
-                oauth_like.add(name)
-            if name == "gemini-cli-cloud-project":
-                oauth_like.add(name)
-
-        functional.update(oauth_like.intersection(set(registered)))
-
-        # Include any dynamically added backends present in __dict__ that have api_key
-        # (used in tests and when users add custom backends not in the registry).
-        for name, cfg in getattr(self, "__dict__", {}).items():
-            if (
-                name == "default_backend"
-                or name.startswith("_")
-                or not isinstance(cfg, BackendConfig)
-            ):
-                continue
-            if cfg.api_key:
-                functional.add(name)
-        return functional
-
-    def __getattr__(self, name: str) -> Any:
-        """Allow accessing backend configs as attributes.
-
-        If an attribute for a backend is missing, create a default
-        BackendConfig instance lazily. This ensures tests and runtime
-        code can access `config.backends.openai` / `config.backends.gemini`
-        even if the registry hasn't been populated yet.
-        """
-        if name == "default_backend":  # Handle default_backend separately
-            # Ensure we use the explicitly set default_backend if available
-            if "default_backend" in self.__dict__:
-                return self.__dict__["default_backend"]
-            # Otherwise fall back to openai
-            return "openai"
-
-        # Check if the attribute exists in __dict__
-        if name in self.__dict__:
-            return cast(BackendConfig, self.__dict__[name])
-
-        # Avoid creating configs for private/internal attributes to maintain security
-        if name.startswith(("_", "__")):
-            raise AttributeError(
-                f"'{self.__class__.__name__}' object has no attribute '{name}'"
-            )
-
-        # Check if we're still initializing (indicated by presence of __dict__ keys
-        # that suggest initialization hasn't completed). Don't create empty configs
-        # during initialization - let the __init__ method handle it.
-        # Only create empty configs after initialization is complete.
-        if not hasattr(self, "_initialization_complete"):
-            # During initialization, raise AttributeError to let __init__ handle it
-            raise AttributeError(
-                f"'{self.__class__.__name__}' object has no attribute '{name}'"
-            )
-
-        # Lazily create a default backend configuration for unknown backends.
-        # This allows accessing backend configs without pre-registration while
-        # maintaining backward compatibility. Created configs are cached for
-        # subsequent access to avoid creating multiple instances.
-        config = BackendConfig()
-        self.__dict__[name] = config
-        return config
-
-    def model_dump(self, **kwargs: Any) -> dict[str, Any]:
-        """Override model_dump to include default_backend and dynamic backends."""
-        dumped: dict[str, Any] = super().model_dump(**kwargs)
-        # Add dynamic backends to the dumped dictionary
-        for backend_name in backend_registry.get_registered_backends():
-            if backend_name in self.__dict__:
-                config: Any = self.__dict__[backend_name]
-                if isinstance(config, BackendConfig):
-                    dumped[backend_name] = config.model_dump()
-        return dumped
-
-    def model_is_functional(self, model_id: str) -> bool:
-        """Check if a model is available in any functional backend."""
-        if ":" not in model_id:
-            return False  # Invalid format
-
-        backend_name, _ = model_id.split(":", 1)
-        return backend_name in self.functional_backends
-
+ 
 
 class AppConfig(DomainModel, IConfig):
     """Complete application configuration."""
@@ -1892,116 +1208,3 @@ class AppConfig(DomainModel, IConfig):
     def get_gcp_project_id(self) -> str | None:
         """Return the GCP Project ID."""
         return self.gcp_project_id
-
-
-def _merge_dicts(d1: dict[str, Any], d2: dict[str, Any]) -> dict[str, Any]:
-    for k, v in d2.items():
-        if k in d1 and isinstance(d1[k], dict) and isinstance(v, dict):
-            _merge_dicts(d1[k], v)
-        else:
-            d1[k] = v
-    return d1
-
-
-def _set_by_path(target: dict[str, Any], path: str, value: Any) -> None:
-    parts = path.split(".")
-    current: dict[str, Any] = target
-    for key in parts[:-1]:
-        current = current.setdefault(key, {})  # type: ignore[assignment]
-    current[parts[-1]] = value
-
-
-def _get_by_path(source: dict[str, Any], path: str) -> Any:
-    parts = path.split(".")
-    current: Any = source
-    for key in parts:
-        if not isinstance(current, dict):
-            return None
-        current = current.get(key)
-    return current
-
-
-def _flatten_dict(data: dict[str, Any]) -> dict[str, Any]:
-    flattened: dict[str, Any] = {}
-
-    def _walk(value: Any, prefix: str) -> None:
-        if isinstance(value, dict):
-            for key, child in value.items():
-                new_prefix = f"{prefix}.{key}" if prefix else key
-                _walk(child, new_prefix)
-        else:
-            flattened[prefix] = value
-
-    _walk(data, "")
-    return flattened
-
-
-def load_config(
-    config_path: str | Path | None = None,
-    *,
-    resolution: ParameterResolution | None = None,
-    environ: Mapping[str, str] | None = None,
-) -> AppConfig:
-    """
-    Load configuration from file and environment.
-
-    Args:
-        config_path: Optional path to configuration file
-
-    Returns:
-        AppConfig instance
-    """
-    env = os.environ if environ is None else environ
-    res = resolution or ParameterResolution()
-
-    config_data: dict[str, Any] = AppConfig().model_dump()
-
-    if config_path:
-        try:
-            import yaml
-
-            path: Path = Path(config_path)
-            if not path.exists():
-                logger.warning(f"Configuration file not found: {config_path}")
-            else:
-                if path.suffix.lower() not in [".yaml", ".yml"]:
-                    raise ValueError(
-                        f"Unsupported configuration file format: {path.suffix}. Use YAML (.yaml/.yml)."
-                    )
-
-                with open(path, encoding="utf-8") as f:
-                    file_config: dict[str, Any] = yaml.safe_load(f) or {}
-
-                from pathlib import Path as _Path
-
-                from src.core.config.semantic_validation import (
-                    validate_config_semantics,
-                )
-                from src.core.config.yaml_validation import validate_yaml_against_schema
-
-                schema_path = (
-                    _Path.cwd() / "config" / "schemas" / "app_config.schema.yaml"
-                )
-                validate_yaml_against_schema(_Path(path), schema_path)
-                validate_config_semantics(file_config, path)
-
-                _merge_dicts(config_data, file_config)
-                origin = str(path)
-                for name, value in _flatten_dict(file_config).items():
-                    res.record(
-                        name,
-                        value,
-                        ParameterSource.CONFIG_FILE,
-                        origin=origin,
-                    )
-        except Exception as exc:  # type: ignore[misc]
-            logger.critical(f"Error loading configuration file: {exc!s}")
-            raise
-
-    env_config = AppConfig.from_env(environ=env, resolution=res)
-    env_dump = env_config.model_dump()
-    for name in res.latest_by_source(ParameterSource.ENVIRONMENT):
-        value = _get_by_path(env_dump, name)
-        _set_by_path(config_data, name, value)
-
-    return AppConfig.model_validate(config_data)
